requirements.txt
dsmr_parser
influxdb_client
click
paho-mqtt
json